Disambiguate "components" and "repositories"
Currently when running "mcp-microservices build -c heka" the
command will try to git-clone the "heka" repo. This does not
make sense as the component specified with the -c option
here is not a Git repository name, it is the name of the
component we want to build ("heka" here).

If the user doesn't want to fetch all the mcp repositories
when using the "build" command then he should use the
"--repositories-names" flag. For example:

    mcp-microservices --repositories-names ms-debian-base,ms-lma build -c heka

Change-Id: I85bc687f816d19a8c6724d955ae8a3e6f5989baf