Dont use unassigned roles during deploy
Right now, we cant create a role and not use it on some host,
We will get an error. I think it's wrong and I culd wish to
create some pre-defined roles, but not use all of them at once.

For example I want to deploy one set of containers at first deploy
and for the second one change 1 role, so I will deploy different
set of containers.

Change-Id: I37fc96a2b479b11c06d2b9dd805f4e5d4ce659db