Make configmaps creation repeatable
If data of configmap was not updated, configmap
by itself should not be updated too. We could benefit
from it by checking only resourceVersion field of
ConfigMap object leaving comporison part to k8s.

It's very useful to use sort_keys flag of json.dumps
for serialization, so format for configs in configmaps
was changed from yaml to json

Change-Id: Ia808a9a886c38b635285c373bcd4f476aad64887