Add primitive support for specifying replicas
It's a minimal support for specifying number of replicas for
services in the ccp config file. It's badly needed to simplify
testing on scale and testing of the running services in HA.
It's not breaking any kind of backward compatibility, you're
now able to specify it:

    replicas:
        keystone: 3
        nova-api: 42

Corresponding verifications was added into the process of
building topology as it's implicitly part of it.

Additionally, there is a very small tests refactoring plus
few new tests for replicas verification.

Change-Id: I4d6b1064ae9698def91aae341ff6105d1415ff66