Fix dependencies resolution for services
Closes-Bug: 1629756

summary: Replace any service dependencies with set of its containers

Take a look on the next case of DSL:

service:
  name: A
  containers:
    - name: c1
    - name: c2
    - name: c3

service:
  name: B
  containers:
    - name: fail
      daemon:
        dependencies:
          - A

When service A will be started, the entrypoints scripts in containters
of A will do reports to etcd that c1,c2 and c3 are done.

So, the service B will wait indefinitely in this case, because the
entrypoint script do report of state of container via its name, not
service name (e.g. A).

Change-Id: I6078bfbd863443aa545048b5f82ceaf8c671f857