Add new script into tools for networking check
This script is an optional tool that can be used to perform basic
diagnostic of networking in K8s cluster. It is meant to be adjusted to
individual needs.
Originally adopted from openstack/fuel-ccp-installer by
kproskurin@mirantis.com.
The reason to keep it in fuel-ccp is to assit user in troubleshooting
if ccp deployment fails (which is a not unusal case at this stage of
development).

Change-Id: I86b671b465d7ed79362d17dcd32c475e1ea6a8cc