Try to detect public eth iface in OpenStack tests
This is an attempt to detect public network interface on a node
from which simple, built-in OpenStack tests will be run.
In case it fails, fallback is provided to hard-coded value.
As before, interace can be set with command-line option.
Autodetection is useful to run these tests without knowing all details
of env deployed beforehand (e.g. which release of Ubuntu it is based
on) - wherever possible this will save us from hardcoding networking
config into our scripts e.g. CI jobs.

Change-Id: I2cecbba95d589960f8e540045a8a3360ee379787