Allow to specify affinity for jobs
This is needed for backup jobs to be run on specific nodes that have
backup volumes mounted.
Jobs can provide a 'topology_key' value that will be looked up in
topology to find which nodes this job can run on. So for backup one
would need to do smth like this:

nodes:
  node1:
    - backup
roles:
  backup:
    - backup

And add "topology_key: backup" to backup jobs.

Change-Id: I3b51b7a957735873b0de098578e1b83c586f111a