Allow 50% unavailable nodes if upgrading DaemonSet kind
If we have something deployed using DaemonSet kind then it's 99%
high coupled to the host or some host capabilities while not providing
any HA or reliability features and it means that there is no reason
to make upgrade of such services one by one.

Good examples are nova-compute, nova-libvirt. There is no real diff
to update all of them at once or one by one (by default). In future
we'll need to implement much more flexible approach for upgrading
such service, as potentially we'll prefer to upgrade 10% of computes,
wait for some period of time to ensure that they are working correct
and only after that upgrade the rest of the compute nodes.

Change-Id: I477f3db48f459fad2753816a82575aa4174c96a6