Add empty ccp-hw-config.yaml file
We include this file in ccp-configs-common.yaml so it should exist. We
generate content of this file in ccp-multi-deploy.sh script but sometimes we
may want to execute e.g. fetch command with this configs, so if we'll not
have this file it will be fail with 'no such file' msg like now our CI.

Change-Id: I64e029f9a8b8406694877ff2fa130462f72461ae