Print logs for failed containers and increase timeout
In k8s 1.5 there is no more default container for printing logs from
pod level. Even if it was it was fail in logic, because we should always be
sure what logs we want to print - mean logs from all containers in pod
not just from one default one.
Additionally increase timeout for deployment.

Change-Id: I84f480dc3f2ff33e1fc600c9a146cd8063b5a5a0