Secret support
Support of k8s secrets is introduced. To create a secret, put
an additional section 'secrets' to the definition
of your service:

secrets:
  - name: my-k8s-secret
    type: "Opaque"
    data:
      "file1": "some content"
      "file2": "another one"

You can also mount a secret as a volume to the specific
place of the container. To do it, insert the next snippet into
'volumes' section:

- name: volume-name
  path: "/path/to/mount"
  type: secret
  secret:
    secretName: "my-k8s-secret"

The mounted secret should already exist - it should be
either created manually or listed in the 'secrets' section.

Change-Id: Ib7ae8f61c3618a673a527a521a6b9b0a47fa366c
Partial-Bug: #1651392
Partial-Bug: #1651394